# docs/openapi_donations.yaml
...
openapi: 3.1.0
info:
  title: GoodPlay Donations API - GOO-15
  description: |
    **GOO-15 Donation Processing Engine** - Complete virtual wallet and donation system.

    ## Features
    - Virtual wallet management with credit conversion
    - Payment processing with multiple providers
    - Donation workflow with compliance checks
    - Batch processing for high-volume operations
    - Financial analytics and reporting
    - Receipt generation and tax compliance
    - Transaction reconciliation

    ## Endpoints Structure
    - `/api/wallet/*` - Virtual wallet operations
    - `/api/donations/*` - Donation processing
    - `/api/conversion-rates/*` - Credit conversion rates
    - `/api/payments/*` - Payment processing
    - `/api/batch/*` - Batch operations (admin)
    - `/api/compliance/*` - Compliance management (admin)
    - `/api/admin/financial/*` - Financial analytics (admin)

  version: 1.0.0
paths:
...
  /api/compliance/checks:
    post:
      tags:
        - Admin - Compliance
      summary: Initiate compliance check
      description: Start AML, KYC, sanctions, or fraud checks for users or transactions
      security:
        - bearerAuth: []
        - adminRole: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - check_type
              properties:
                user_id:
                  type: string
                check_type:
                  type: string
                  enum: [aml, kyc, sanctions, pep, fraud]
                transaction_id:
                  type: string
                reason:
                  type: string
      responses:
        '200':
          description: Compliance check initiated successfully
  /api/compliance/checks/{check_id}/review:
    post:
      tags:
        - Admin - Compliance
      summary: Review compliance check
      description: Manually review and approve/reject compliance checks
      security:
        - bearerAuth: []
        - adminRole: []
      parameters:
        - name: check_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - decision
              properties:
                decision:
                  type: string
                  enum: [approve, reject, escalate]
                review_notes:
                  type: string
      responses:
        '200':
          description: Compliance check reviewed successfully

  # Admin Financial Analytics Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminRole:
      type: apiKey
      in: header
      name: X-Admin-Role
...